refactor: move GroupEvent to a new file
